# Conflicts:
#	openstudiocore/resources/model/OpenStudio.idd
#	openstudiocore/src/model/AirLoopHVACOutdoorAirSystem_Impl.hpp
#	openstudiocore/src/model/AirLoopHVACZoneMixer.hpp
#	openstudiocore/src/model/AirTerminalSingleDuctConstantVolumeReheat.cpp
#	openstudiocore/src/model/AirTerminalSingleDuctConstantVolumeReheat.hpp
#	openstudiocore/src/model/AirTerminalSingleDuctConstantVolumeReheat_Impl.hpp
#	openstudiocore/src/model/AirTerminalSingleDuctVAVReheat.cpp
#	openstudiocore/src/model/CoilCoolingDXSingleSpeed.cpp
#	openstudiocore/src/model/CoilCoolingDXSingleSpeed.hpp
#	openstudiocore/src/model/CoilCoolingDXTwoStageWithHumidityControlMode.cpp
#	openstudiocore/src/model/CoilCoolingDXTwoStageWithHumidityControlMode_Impl.hpp
#	openstudiocore/src/model/CoilCoolingWater.cpp
#	openstudiocore/src/model/CoilCoolingWater.hpp
#	openstudiocore/src/model/CoilCoolingWater_Impl.hpp
#	openstudiocore/src/model/CoilHeatingDXMultiSpeed.cpp
#	openstudiocore/src/model/CoilHeatingDXMultiSpeed.hpp
#	openstudiocore/src/model/CoilHeatingDXSingleSpeed.cpp
#	openstudiocore/src/model/CoilHeatingDXSingleSpeed_Impl.hpp
#	openstudiocore/src/model/CoilHeatingElectric.cpp
#	openstudiocore/src/model/CoilHeatingElectric.hpp
#	openstudiocore/src/model/CoilHeatingElectric_Impl.hpp
#	openstudiocore/src/model/CoilHeatingGas.cpp
#	openstudiocore/src/model/CoilHeatingGas.hpp
#	openstudiocore/src/model/CoilHeatingGas_Impl.hpp
#	openstudiocore/src/model/CoilHeatingWater.cpp
#	openstudiocore/src/model/CoilHeatingWater_Impl.hpp
#	openstudiocore/src/model/ControllerOutdoorAir.cpp
#	openstudiocore/src/model/ControllerOutdoorAir_Impl.hpp
#	openstudiocore/src/model/FanConstantVolume.cpp
#	openstudiocore/src/model/FanConstantVolume.hpp
#	openstudiocore/src/model/FanConstantVolume_Impl.hpp
#	openstudiocore/src/model/FanOnOff.cpp
#	openstudiocore/src/model/FanOnOff.hpp
#	openstudiocore/src/model/FanVariableVolume.cpp
#	openstudiocore/src/model/FanVariableVolume.hpp
#	openstudiocore/src/model/FanVariableVolume_Impl.hpp
#	openstudiocore/src/model/HeatExchangerAirToAirSensibleAndLatent.cpp
#	openstudiocore/src/model/HeatExchangerAirToAirSensibleAndLatent_Impl.hpp
#	openstudiocore/src/model/Model.i
#	openstudiocore/src/model/ModelGeometry.i
#	openstudiocore/src/model/SubSurface_Impl.hpp
#	openstudiocore/src/model/Surface.cpp
#	openstudiocore/src/model/Surface.hpp
#	openstudiocore/src/model/Surface_Impl.hpp

+++ openstudiocore/src/utilities/data/Vector.cpp
@@ -1,5 +1,5 @@
 /***********************************************************************************************************************
- *  OpenStudio(R), Copyright (c) 2008-2017, Alliance for Sustainable Energy, LLC. All rights reserved.
+ *  OpenStudio(R), Copyright (c) 2008-2018, Alliance for Sustainable Energy, LLC. All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
  *  following conditions are met:
@@ -38,7 +38,7 @@
 namespace openstudio{
 
   Vector createVector(const std::vector<double>& values) {
-    unsigned n = values.size();
+    size_t n = values.size();
     Vector result(n);
     for (unsigned i = 0; i < n; ++i) {
       result[i] = values[i];
@@ -47,7 +47,7 @@ namespace openstudio{
   }
 
   Vector createVector(const std::vector<long>& values) {
-    unsigned n = values.size();
+    size_t n = values.size();
     Vector result(n);
     for (unsigned i = 0; i < n; ++i) {
       result[i] = (double)(values[i]);
@@ -56,9 +56,9 @@ namespace openstudio{
   }
 
   std::vector<double> toStandardVector(const Vector& values) {
-    unsigned n = values.size();
+    size_t n = values.size();
     std::vector<double> result(n);
-    for (unsigned i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
       result[i] = values[i];
     }
     return result;
@@ -82,7 +82,7 @@ namespace openstudio{
   /// assumes that x is strictly increasing
   InterpInfo interpInfo(const Vector& x, double xi)
   {
-    unsigned N = x.size();
+    size_t N = x.size();
 
     InterpInfo result;
 
@@ -100,21 +100,21 @@ namespace openstudio{
       result.extrapolated = false;
       Vector::const_iterator begin = x.begin();
       Vector::const_iterator it = std::lower_bound(begin, x.end(), xi);
-      
+
       result.ia = (unsigned)(it-begin-1);
       result.ib = (unsigned)(it-begin);
       result.wa = (x(result.ib)-xi)/(x(result.ib)-x(result.ia));
       result.wb = (xi-x(result.ia))/(x(result.ib)-x(result.ia));
     }
-      
+
     return result;
   }
 
   /// linear interpolation of the function y = f(x) at point xi
   /// assumes that x is strictly increasing
   double interp(const Vector& x, const Vector& y, double xi, InterpMethod interpMethod, ExtrapMethod extrapMethod){
-    
-    unsigned N = x.size();
+
+    size_t N = x.size();
 
     double result = 0.0;
 
@@ -162,8 +162,8 @@ namespace openstudio{
   /// linear interpolation of the function y = f(x) at points xi
   /// assumes that x is strictly increasing
   Vector interp(const Vector& x, const Vector& y, const Vector& xi, InterpMethod interpMethod, ExtrapMethod extrapMethod){
- 
-    unsigned N = x.size();
+
+    size_t N = x.size();
 
     Vector result(N);
 
@@ -183,7 +183,7 @@ namespace openstudio{
   {
     // seed random number generator
     static std::minstd_rand generator(42u);
-    
+
     // define distribution
     boost::uniform_real<> dist(a,b);
 
@@ -202,7 +202,7 @@ namespace openstudio{
   Vector linspace(double a, double b, unsigned N)
   {
     double delta = (b-a)/(N-1);
-    
+
     Vector result(N);
     for (unsigned n = 0; n < N; ++n){
       result(n) = a + n*delta;
@@ -240,7 +240,7 @@ namespace openstudio{
   /// take the natural logarithm of elements of a Vector
   Vector log(const Vector& x)
   {
-    unsigned N = x.size();
+    size_t N = x.size();
     Vector result(N);
     for (unsigned i = 0; i < N; ++i){
       result(i) = std::log(x(i));
@@ -252,7 +252,7 @@ namespace openstudio{
   Vector log(const Vector& x, double base)
   {
     double logBase = std::log(base);
-    unsigned N = x.size();
+    size_t N = x.size();
     Vector result(N);
     for (unsigned i = 0; i < N; ++i){
       result(i) = std::log(x(i)) / logBase;
@@ -308,7 +308,7 @@ namespace openstudio{
   double mean(const Vector& vector)
   {
     double avg = 0;
-    unsigned N = vector.size();
+    size_t N = vector.size();
     if (N > 0){
       avg = sum(vector) / N;
     }
@@ -319,7 +319,7 @@ namespace openstudio{
   double variance(const Vector& vector)
   {
     double result = 0;
-    unsigned N = vector.size();
+    size_t N = vector.size();
     if (N > 0)
     {
       double sumSquares = pow(boost::numeric::ublas::norm_2(vector), 2.0);
@@ -334,8 +334,8 @@ namespace openstudio{
     return sqrt(variance(vector));
   }
 
-  std::function<double (const Vector&)> sumVectorFunctor() { 
-    return std::function<double (const Vector&)> (&sum); 
+  std::function<double (const Vector&)> sumVectorFunctor() {
+    return std::function<double (const Vector&)> (&sum);
   }
 
   std::function<double (const Vector&)> maximumVectorFunctor() {
