+++ openstudiocore/src/utilities/geometry/Plane.cpp
@@ -1,5 +1,5 @@
 /***********************************************************************************************************************
- *  OpenStudio(R), Copyright (c) 2008-2017, Alliance for Sustainable Energy, LLC. All rights reserved.
+ *  OpenStudio(R), Copyright (c) 2008-2018, Alliance for Sustainable Energy, LLC. All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
  *  following conditions are met:
@@ -64,11 +64,11 @@ namespace openstudio{
 
   double det3x3(const Matrix& m)
   {
-    double det = m(0,0)*m(1,1)*m(2,2) + 
+    double det = m(0,0)*m(1,1)*m(2,2) +
                  m(0,1)*m(1,2)*m(2,0) +
                  m(0,2)*m(1,0)*m(2,1) -
                  m(0,2)*m(1,1)*m(2,0) -
-                 m(0,1)*m(1,0)*m(2,2) - 
+                 m(0,1)*m(1,0)*m(2,2) -
                  m(0,0)*m(1,2)*m(2,1);
     return det;
   }
@@ -76,7 +76,7 @@ namespace openstudio{
   Plane::Plane(const std::vector<Point3d>& points)
     : m_a(0.0), m_b(0.0), m_c(0.0), m_d(0.0)
   {
-    unsigned N = points.size();
+    size_t N = points.size();
     if (N < 3){
       LOG_AND_THROW("Cannot compute plane with fewer than three points");
     }else if (N == 3){
@@ -110,15 +110,15 @@ namespace openstudio{
         Matrix At(3,N);
         Vector b(N);
         for (unsigned i = 0; i < N; ++i){
-          A(i,0) = points[i].x() - points[0].x(); 
+          A(i,0) = points[i].x() - points[0].x();
           A(i,1) = points[i].y() - points[0].y();
           A(i,2) = 1.0;
-          At(0,i) = A(i,0); 
+          At(0,i) = A(i,0);
           At(1,i) = A(i,1);
           At(2,i) = A(i,2);
           b[i] = -(points[i].z() - points[0].z());
         }
-      
+
         Matrix AtA = prod(At, A);
         double det = det3x3(AtA); // always positive for A'*A
         if (det > maxDet){
@@ -152,15 +152,15 @@ namespace openstudio{
         Matrix At(3,N);
         Vector b(N);
         for (unsigned i = 0; i < N; ++i){
-          A(i,0) = points[i].x() - points[0].x(); 
+          A(i,0) = points[i].x() - points[0].x();
           A(i,1) = points[i].z() - points[0].z();
           A(i,2) = 1.0;
-          At(0,i) = A(i,0); 
+          At(0,i) = A(i,0);
           At(1,i) = A(i,1);
           At(2,i) = A(i,2);
           b[i] = -(points[i].y() - points[0].y());
         }
-      
+
         Matrix AtA = prod(At, A);
         double det = det3x3(AtA); // always positive for A'*A
         if (det > maxDet){
@@ -194,15 +194,15 @@ namespace openstudio{
         Matrix At(3,N);
         Vector b(N);
         for (unsigned i = 0; i < N; ++i){
-          A(i,0) = points[i].y() - points[0].y(); 
+          A(i,0) = points[i].y() - points[0].y();
           A(i,1) = points[i].z() - points[0].z();
           A(i,2) = 1.0;
-          At(0,i) = A(i,0); 
+          At(0,i) = A(i,0);
           At(1,i) = A(i,1);
           At(2,i) = A(i,2);
           b[i] = -(points[i].x() - points[0].x());
         }
-      
+
         Matrix AtA = prod(At, A);
         double det = det3x3(AtA); // always positive for A'*A
         if (det > maxDet){
